Code tidy

+++ src/util/number.rs
@@ -39,7 +39,7 @@ impl Number {
 
     pub fn to_i16(&self) -> Option<i16> {
         match *self {
-            Number::U8(n) => Some(n as i8 as i16),
+            Number::U8(n) => Some(n as i16),
             Number::I16(n) => Some(n),
             Number::I32(n) => n.to_i16(),
             Number::I64(n) => n.to_i16(),
@@ -50,7 +50,7 @@ impl Number {
 
     pub fn to_i32(&self) -> Option<i32> {
         match *self {
-            Number::U8(n) => Some(n as i8 as i32),
+            Number::U8(n) => Some(n as i32),
             Number::I16(n) => Some(n as i32),
             Number::I32(n) => Some(n),
             Number::I64(n) => n.to_i32(),
@@ -61,7 +61,7 @@ impl Number {
 
     pub fn to_i64(&self) -> Option<i64> {
         match *self {
-            Number::U8(n) => Some(n as i8 as i64),
+            Number::U8(n) => Some(n as i64),
             Number::I16(n) => Some(n as i64),
             Number::I32(n) => Some(n as i64),
             Number::I64(n) => Some(n),
@@ -72,7 +72,7 @@ impl Number {
 
     pub fn to_f32(&self) -> Option<f32> {
         match *self {
-            Number::U8(n) => (n as i8 as i32).to_f32(),
+            Number::U8(n) => (n as i32).to_f32(),
             Number::I16(n) => (n as i32).to_f32(),
             Number::I32(n) => n.to_f32(),
             Number::I64(n) => n.to_f32(),
@@ -83,7 +83,7 @@ impl Number {
 
     pub fn to_f64(&self) -> Option<f64> {
         match *self {
-            Number::U8(n) => (n as i8 as i32).to_f64(),
+            Number::U8(n) => (n as i32).to_f64(),
             Number::I16(n) => (n as i32).to_f64(),
             Number::I32(n) => n.to_f64(),
             Number::I64(n) => n.to_f64(),
